+++ crates/amaru-kernel/src/lib.rs
@@ -49,9 +49,9 @@ pub use data_structures::{
 pub mod cbor {
     pub use amaru_minicbor_extra::{
         TAG_MAP_259, TAG_SET_258, WithSize, allow_tag, check_tagged_array_length, collect_array_item_bytes,
-        collect_map_value_bytes, decode_break, decode_bytes, decode_string, expect_tag, from_cbor,
-        from_cbor_no_leftovers, from_cbor_no_leftovers_with, heterogeneous_array, heterogeneous_map, lazy,
-        missing_field, tee, to_cbor, to_cbor_with, unexpected_field,
+        collect_map_value_bytes, count_bytes, decode_break, decode_bytes, decode_string, encode_variable_length_map,
+        expect_tag, from_cbor, from_cbor_no_leftovers, from_cbor_no_leftovers_with, heterogeneous_array,
+        heterogeneous_map, lazy, missing_field, tee, to_cbor, to_cbor_with, unexpected_field,
     };
     pub use minicbor::{
         CborLen, Decode, Decoder, Encode, Encoder, bytes,
@@ -118,7 +118,7 @@ pub use cardano::{
     max_string::MaxString128,
     memoized::{
         self, BorrowedScript, MemoizedDatum, MemoizedNativeScript, MemoizedPlutusData, MemoizedScript,
-        MemoizedTransactionOutput, MemoizedValue, deserialize_script, serialize_memoized_script, serialize_script,
+        MemoizedTransactionOutput, deserialize_script, serialize_memoized_script, serialize_script,
     },
     metadatum::Metadatum,
     multiasset::{self, Multiasset},