+++ crates/amaru-plutus/src/script_context/v2.rs
@@ -66,7 +66,7 @@ impl ToPlutusData<2> for OutputReference<'_> {
 
 impl ToPlutusData<2> for MemoizedTransactionOutput {
     fn to_plutus_data(&self) -> Result<PlutusData, PlutusDataError> {
-        constr_v2!(0, [self.address, self.value.as_ref(), self.datum, self.script])
+        constr_v2!(0, [self.address, self.value, self.datum, self.script])
     }
 }